requirements.txt
# Aviator Project Requirements
# Python 3.8+ required

# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
Pillow>=9.0.0
pytesseract>=0.3.9

# Machine Learning
scikit-learn>=1.0.0
joblib>=1.1.0

# GUI Automation
pyautogui>=0.9.53
pynput>=1.7.6
keyboard>=0.13.5
mouse>=0.7.1

# Screen capture
mss>=7.0.0
opencv-python>=4.5.0

# Database
# sqlite3 is built-in

# Utilities
psutil>=5.9.0
colorama>=0.4.4

# Development tools (optional)
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950

# Performance monitoring (optional)
matplotlib>=3.5.0
seaborn>=0.11.2

# Windows specific (install only on Windows)
# pywin32>=303 ; platform_system == "Windows"